move metrics into to metrics.py and remove id from title
