+++ docs/source/generate_metric_list.py
@@ -4,7 +4,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from parlai.core.logs import METRICS_DISPLAY_DATA
+from parlai.core.metrics import METRICS_DISPLAY_DATA
 
 
 fout = open('metric_list.inc', 'w')
